requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
clickclick==20.10.2
connexion==2.12.0
distlib==0.3.4
docopt==0.6.2
filelock==3.6.0
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
pipenv==2022.1.8
platformdirs==2.5.1
pyrsistent==0.18.1
python-dotenv==0.19.2
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
stomp.py==7.0.0
structlog==20.1.0
urllib3==1.26.8
virtualenv==20.13.3
virtualenv-clone==0.5.7
Werkzeug==2.0.3
dependency-injector==4.38.0
PyJWT[crypto]==2.3.0
jcs==0.2.0
pymongo==4.1.1
tenacity==8.0.1
cryptography==37.0.2
GitPython==3.1.27
py-healthcheck==1.10.1
